requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.54
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==3.0.1
PyYAML==6.0.2
